% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsignal.R
\name{falsignal}
\alias{falsignal}
\title{Find false signal}
\usage{
falsignal(l, cutoffvar = 0.03, cutoffswt = 1, cutofflong = NULL,
  cutoffmr = 0.95)
}
\arguments{
\item{l}{A list generated by \code{\link{makefeature}}.}

\item{cutoffvar}{Metabolites whose variance >= cutoffvar tend to be false signal.}

\item{cutoffswt}{Metabolites whose number of switches >= cutoffswt tend to be false signal.}

\item{cutofflong}{Metabolites whose longest block < cutofflong tend to be false signal.}

\item{cutoffmr}{Metabolites whose mean missing rate > cutoffmr tend to be false signal}
}
\value{
A list contains name lists of predicted false signal.
\itemize{
\item mbnames: name list of metabolites.
\item wsize: window size.
\item ssize: slide size.
\item defswitch: definition of switch.
\item mrate: missing rate matrix. Each cell of the matrix is a missing rate of a window.
\item variance: missing rate variance for each metabolite.
\item nswitches: number of switches for each metabolite.
\item longestblock: number of windows of the longest block for each metabolite.
\item boolbad1: bool vector of type 1 false signal.
\item boolbad2: bool vector of type 2 false signal.
}
}
\author{
Liu Cao
}

