% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefeature.R
\name{makefeature}
\alias{makefeature}
\title{Generate missing rate matrix and extract 3 features characterizing the missing rate pattern}
\usage{
makefeature(data, wsize = 100, ssize = 0.5, defswitch = 0.2)
}
\arguments{
\item{data}{A data frame of metabolites signal. Rows are metabolites and columns are samples in injection order. The data frame can be created by \code{\link{read.metfile}}.}

\item{wsize}{Window size. Default is 100.}

\item{ssize}{Slide size. Default is 0.5.}

\item{defswitch}{Definition of a switch. If the absolute missing rate difference of adjacent windows is greater than defswitch, it will be defined as a switch. Default is 0.2.}
}
\value{
A list contains features generated from the metabolites signal matrix. The list contains the following entries:
\itemize{
\item mbnames: A vector. Name list of metabolites.
\item snames: A vector. Name list of samples.
\item wsize: A scalar. Window size.
\item ssize: A scalar. Slide size.
\item defswitch: A scalar. Definition of switch.
\item mrate: Missing rate matrix. Each cell of the matrix is a missing rate of a window.
\item variance: A vector. Missing rate variance for each metabolite.
\item nswitches: A vector. Number of switches for each metabolite.
\item longestblock: A vector. Number of windows of the longest block for each metabolite.
}
}
\description{
\code{makefeature} generates the missing rate matrix from the data frame of metabolomics signal (generated by \code{\link{read.metfile}}), according to specified window size and slide size.
With the missing rate matrix, \code{makefeature} gives 3 features characterizing the missing rates pattern of the metabolites, including variance of missing rate, number of switches and longest block.
}
\author{
Liu Cao
}
\seealso{
See \code{\link{read.metfile}} for how to generate the the input metabolites signal data frame.
}

