% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mrate}
\alias{plot_mrate}
\title{Plot missing rate}
\usage{
plot_mrate(bds, type = "1", mname)
}
\arguments{
\item{bds}{A list returned by \code{\link{falsignal}}.}

\item{mname}{The name of metabolite if method="single".}

\item{method}{"all" plots the missing rate of all metabolites. "bad1" only plots the missing rate of type 1 bad signal. "bad2" only plots the missing rate of type 2 bad signal. "single" only plots missing rate for a given metabolite.}
}
\value{
none
}
\author{
Liu Cao
}
\seealso{
See \code{\link{falsignal}} for how to generate the input list \code{bds}
}

