% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot QC figures}
\usage{
plot_qc(bds)
}
\arguments{
\item{bds}{A list returned by falsignal().}

\item{method}{"all" plots the missing rate of all metabolites. "bad1" only plots the missing rate of type 1 bad signal. "bad2" only plots the missing rate of type 2 bad signal. "single" only plots missing rate for a given metabolite.}

\item{mname}{The name of metabolite if method="single".}
}
\value{
none
}
\description{
\code{\link{plot_qc}} plot a set of quality control figures that summarize the distribution of
 missing pattern and features of predicted false metabolites and good metabolites.
}
\author{
Liu Cao
}

