% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.metfile}
\alias{write.metfile}
\title{Write metabolomics data}
\usage{
write.metfile(data, falsig, type = "1", trans = TRUE)
}
\arguments{
\item{data}{Data frame generated by \code{\link{read.metfile}}}

\item{falsig}{The list generated by \code{\link{falsignal}}}

\item{type}{type=1 outputs type I false signal. type=2 outputs type II false signal. Default is 1.}

\item{trans}{whether to transponse the table. Default is TRUE, or with rows of samples and columns of metabolites.}
}
\value{
None
}
\description{
\code{write.metfile} output csv file of false or good metabolites signal according to type I criterion or type II criterion.
}
\author{
Liu Cao
}

